requirements.txt
asgiref==3.11.0
boto3==1.42.46
botocore==1.42.46
certifi==2026.1.4
charset-normalizer==3.4.4
cloudinary==1.44.1
dj-database-url==3.0.1
Django==6.0
django-cloudinary-storage==0.3.0
django-humanize==0.1.2
django-storages==1.14.6
gunicorn==23.0.0
humanize==4.15.0
idna==3.11
jmespath==1.1.0
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.4
stripe==14.2.0
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.11.0
